+++ src/game/tests.rs
@@ -512,6 +512,7 @@ mod vintage_cube_shieldbreaker;
 mod vintage_cube_shifting_woodland;
 mod vintage_cube_show_and_tell;
 mod vintage_cube_six;
+mod vintage_cube_skyclave_apparition;
 mod vintage_cube_smugglers_copter;
 mod vintage_cube_spells;
 mod vintage_cube_staff_of_the_storyteller;